* common-src/security-util.c: Fix for compiler warning
git-svn-id: https://svn.code.sf.net/p/amanda/code/amanda/trunk@7206 a8d146d6-cc15-0410-8900-af154a0219e0

+++ common-src/security-util.c
@@ -2118,7 +2118,7 @@ sec_tcp_conn_put(
 	    count--;
 	}
 	if (pid == 0) {
-	    g_debug("sending SIGTERM to pid: %d", rc->pid);
+	    g_debug("sending SIGTERM to pid: %ld", (long)rc->pid);
 	    kill(rc->pid, SIGTERM);
 	    pid = waitpid(rc->pid, &status, WNOHANG);
 	    count = 50;
